Fixed Typing Issues
